feat(delivery/order-api): Add GET endpoints for delivered orders
Adds dedicated API routes for clients and administrators to retrieve a list of orders that have been successfully delivered.

+++ src/main/java/org/upc/cobox/fleet/domain/model/commands/CreateFleetCommand.java
@@ -2,4 +2,6 @@ package org.upc.cobox.fleet.domain.model.commands;
 
 import org.upc.cobox.fleet.domain.model.valueobjects.Placa;
 
-public record CreateFleetCommand(Placa placa, String marca, String modelo, Integer capacidadKg) {}
+import java.math.BigDecimal;
+
+public record CreateFleetCommand(Placa placa, String marca, String modelo, BigDecimal capacidadKg) {}
